LW-617 Reorganize account components
