+++ src/components/App/index.js
@@ -5,9 +5,9 @@ import { createStore, applyMiddleware, compose } from 'redux'
 
 import PageWrapper from '../../components/PageWrapper'
 import Home from '../../components/Home'
-import PersonalInfo from '../../components/PersonalInfo'
-import Courses from '../../components/Courses'
-import Settings from '../../components/Settings'
+import PersonalInfo from '../../components/Account/PersonalInfo'
+import Courses from '../../components/Account/Courses'
+import Settings from '../../components/Account/Settings'
 import Hours from '../../components/Hours/Page'
 import ChatPage from '../../components/ChatPage'
 import ContentfulPage from '../../components/Contentful/Page'
@@ -27,7 +27,7 @@ import rootReducers from '../../reducers'
 import thunkMiddleware from 'redux-thunk'
 import Rewrite from './Rewrite'
 import EmbeddableHours from '../Embeddable/Hours'
-import CirculationHistory from '../LoanResources/CirculationHistory'
+import CirculationHistory from '../Account/CirculationHistory'
 
 import NotFound from '../../components/Messages/NotFound'
